test(qa): isolate provider lane selection regressions
